nothing

+++ src/pages/edit-folder/edit-folder.ts
@@ -3,6 +3,7 @@ import { IonicPage, NavController, NavParams } from 'ionic-angular';
 import { SQLite, SQLiteObject } from '@ionic-native/sqlite';
 import { AuthService } from '../../service/auth.service';
 import { File } from '@ionic-native/file';
+import { Toast } from '@ionic-native/toast';
 
 @IonicPage()
 @Component({
@@ -11,7 +12,7 @@ import { File } from '@ionic-native/file';
 })
 export class EditFolderPage {
 
-	folder = { folderid:0, name:"", date:"", type:"", display:"yes" };
+	folder = { folderid:0, name:"" };
 	data = this.auth.getEmail();
 	dataPhone = this.auth.getPhone();
 	foldername = this.navParams.get('foldername');
@@ -22,7 +23,8 @@ export class EditFolderPage {
 		public navParams: NavParams,
 		private sqlite: SQLite,
 		private auth: AuthService,
-		private file: File
+		private file: File,
+		private toast: Toast,
 		) {
 		this.getCurrentFolder(navParams.get("folderid"));
 	}
@@ -39,10 +41,7 @@ export class EditFolderPage {
 				.then(res => {
 					if(res.rows.length > 0) {
 						this.folder.folderid = res.rows.item(0).folderid;
-						this.folder.name = res.rows.item(0).name;
-						this.folder.date = res.rows.item(0).date;
-						this.folder.type = res.rows.item(0).type;
-						this.folder.display = res.rows.item(0).display;						
+						this.folder.name = res.rows.item(0).name;					
 					}		
 				}).catch(e => console.log('Select nothing from Folder table: ' + e.message));
 			}).catch(e => console.log('SQLite didn\'t create SQLite: ' + e.message));
@@ -60,10 +59,7 @@ export class EditFolderPage {
 				.then(res => {
 					if(res.rows.length > 0) {
 						this.folder.folderid = res.rows.item(0).folderid;
-						this.folder.name = res.rows.item(0).name;
-						this.folder.date = res.rows.item(0).date;
-						this.folder.type = res.rows.item(0).type;
-						this.folder.display = res.rows.item(0).display;						
+						this.folder.name = res.rows.item(0).name;					
 					}		
 				}).catch(e => console.log('Select nothing from Folder table: ' + e.message));
 			}).catch(e => console.log('SQLite didn\'t create SQLite: ' + e.message));
@@ -80,10 +76,15 @@ export class EditFolderPage {
 				name: nameDB,
 				location: 'default'
 			}).then((db: SQLiteObject) => {
-				db.executeSql('UPDATE folder SET name=?,date=?,type=?,display=? WHERE folderid=?',[this.folder.name,this.folder.date,this.folder.type,this.folder.display,this.folder.folderid]).then(res => {
-					this.file.copyDir(this.path, oldName, this.path, name).catch(e => console.log('Folder didn\'t copy: ' + e.message));
-					this.file.removeRecursively(this.path, oldName).catch(e => console.log('Folder didn\'t remove: ' + e.message));
-					this.navCtrl.popToRoot();
+				db.executeSql('UPDATE folder SET name=? WHERE folderid=?',[this.folder.name,this.folder.folderid]).then(res => {
+					if (this.folder.name!=null) {
+						this.file.copyDir(this.path, oldName, this.path, name).catch(e => console.log('Folder didn\'t copy: ' + e.message));
+						this.file.removeRecursively(this.path, oldName).catch(e => console.log('Folder didn\'t remove: ' + e.message));
+						this.navCtrl.popToRoot();
+					}
+					else {
+						this.toast.show('Tên thư mục không được để trống', '5000', 'center').subscribe(toast => console.log(toast))
+					}
 				}).catch(e => console.log('Folder didn\'t update: ' + e.message));					  								
 			}).catch(e => console.log('SQLite didn\'t create: ' + e.message));	
 		}
@@ -97,10 +98,15 @@ export class EditFolderPage {
 				name: nameDB,
 				location: 'default'
 			}).then((db: SQLiteObject) => {			
-				db.executeSql('UPDATE folder SET name=?,date=?,type=?,display=? WHERE folderid=?',[this.folder.name,this.folder.date,this.folder.type,this.folder.display,this.folder.folderid]).then(res => {
-					this.file.copyDir(this.path, oldName, this.path, name).catch(e => console.log('Folder didn\'t copy: ' + e.message));
-					this.file.removeRecursively(this.path, oldName).catch(e => console.log('Folder didn\'t remove: ' + e.message));
-					this.navCtrl.popToRoot();
+				db.executeSql('UPDATE folder SET name=? WHERE folderid=?',[this.folder.name,this.folder.folderid]).then(res => {
+					if (this.folder.name!=null) {
+						this.file.copyDir(this.path, oldName, this.path, name).catch(e => console.log('Folder didn\'t copy: ' + e.message));
+						this.file.removeRecursively(this.path, oldName).catch(e => console.log('Folder didn\'t remove: ' + e.message));
+						this.navCtrl.popToRoot();
+					}
+					else {
+						this.toast.show('Tên thư mục không được để trống', '5000', 'center').subscribe(toast => console.log(toast))
+					}
 				}).catch(e => console.log('Folder didn\'t update: ' + e.message));					  								
 			}).catch(e => console.log('SQLite didn\'t create: ' + e.message));	
 		}
